+++ gatenlp/processing/gazetteer/tokengazetteer.py
@@ -5,12 +5,18 @@ gazetteer lists.
 """
 
 import os
+from typing import Union, Dict, Optional, Callable, List
 from collections import defaultdict
-
 from recordclass import structclass
-from gatenlp.document import Document
+
+from gatenlp.document import Document, Annotation
 from gatenlp.utils import init_logger
+from gatenlp.processing.annotator import Annotator
 from gatenlp.processing.gazetteer.base import GazetteerAnnotator
+
+# TODO: better handling/support for separator annotations: this would add complexity but allow that a sequence
+#   of annotations is only matched if there is a/several? separator annotation between each of those annotations.
+#   Could also require this only if there is a separator in the gazetteer sequence (e.g. indicated by a None element)
 
 
 # NOTE! this was origiannl a @dataclass(unsafe_hash=True, order=True)
@@ -74,34 +80,33 @@ def tokentext_getter(token, doc=None, feature=None):
 class TokenGazetteer(GazetteerAnnotator):
     def __init__(
         self,
-        source,
-        fmt="gate-def",
+        source: Union[List, str, None] = None,
+        source_fmt: str = "gate-def",
         source_sep="\t",
         source_encoding="UTF-8",
         # cache_source=None,   # TODO
-        tokenizer=None,
-        all=False,
-        skip=True,
-        outset="",
-        outtype="Lookup",
-        annset="",
-        tokentype="Token",
+        source_tokenizer: Union[None, Annotator, Callable] = None,
+        longest_only: bool = False,
+        skip_longest: bool = False,
+        outset_name: str = "",
+        ann_type: str = "Lookup",
+        annset_name: str = "",
+        token_type: str = "Token",
         feature=None,
-        septype=None,
-        splittype=None,
-        withintype=None,
-        mapfunc=None,
-        ignorefunc=None,
-        getterfunc=None,
-        listfeatures=None,
-        listtype=None,
+        split_type: Optional[str] = None,
+        within_type: Optional[str] = None,
+        mapfunc: Optional[Callable] = None,
+        ignorefunc: Optional[Callable] = None,
+        getterfunc: Optional[Callable] = None,
+        list_features: Optional[Dict] = None,
+        list_type: Optional[str] = None,
     ):
         """
 
         Args:
             source: where to load the gazetteer from. What is actually expected here depends on the fmt
-              parameter.
-            fmt: defines what is expected as the format and/or content of the source parameter. One of:
+              parameter. If none, nothing is loaded
+            source_fmt: defines what is expected as the format and/or content of the source parameter. One of:
                *  "gate-def" (default): the path to a GATE-style "def" file.
                   See https://gate.ac.uk/userguide/chap:gazetteers
                * "gazlist": a list of tuples or lists where the first element of the tuple/list
@@ -110,46 +115,48 @@ class TokenGazetteer(GazetteerAnnotator):
                   specified with the listfeatures parameter and a list type as specified with the listtype parameter.
             source_sep: the field separator to use for some source formats (default: tab character)
             source_encoding: the encoding to use for some source formats (default: UTF-8)
+            source_tokenizer: if not None, an annotator, that creates annotations of type "Token" in the default
+                annotation set. If this is None, then when loading string gazetteer entries, they are tokenized by
+                splitting on whitespace (as defined by Python str.split())
             feature: the feature name to use to get the string for each token. If the corresponding feature
                 in the token does not exist, is None or is the empty string, the Token is completely ignored.
                 If the feature parameter is None, use the document string covered by the token.
-            all: return all matches, if False only return longest matches
-            skip: skip forward over longest match (do not return contained/overlapping matches)
-            annset: the set where the tokens to match should come from
-            outset: the set where the new annotations are added
-            outtype: the annotation type of the annotations to create, unless a type is given for the gazetteer
-               entry or for the gazetteer list.
-            tokentype: the annotation type of the token annotations
-            septype: the annotation type of separator annotations (NOT YET USED/IMPLEMENTED!)
-            splittype: the annotation type of any split annotations which will end any ongoing match
-            withintype: only matches fully within annotations of this type will be made
-            mapfunc: a function that maps the original string extracted for each token to the actual string to use.
-            ignorefunc: a function which given the mapped token string decides if the token should be ignored
-              (not added to the gazetteer list, not considered in the document when matching)
-            getterfunc: a function which, given a token annotation, retrieves the string. If there is mapfunc, the
-              retrieved string is then still run through the mapfunc. The getterfunc must accept the token and
-              an optional document as parameters.
-            listfeatures: a dictionary of features common to the whole list or None. If what gets loaded specifies
-              its own list features, this is getting ignored.
-            listtype: the output annotation type to use for the list, ignored if the input format specifies this
-              on its own. If the input does not specify this on its own and this is not None, then it takes
-              precedence over outtype for the data loaded from source.
+            longest_only: if True, only returns the longest match at each matching position, otherwise returns all
+                matches.
+            skip_longest: skip forward over longest match (do not return contained/overlapping matches)
+            annset_name: the set where the tokens to match should come from
+            outset_name: the set where the new annotations are added
+            ann_type: the annotation type of the annotations to create, unless a type is given for the gazetteer
+                entry or for the gazetteer list.
+            token_type: the annotation type of the token annotations
+            split_type: the annotation type of any split annotations which will end any ongoing match
+            within_type: only matches fully within annotations of this type will be made
+            mapfunc: a callable that maps the original string extracted for each token to the actual string to use.
+            ignorefunc: a callable which given the mapped token string decides if the token should be ignored
+                (not added to the gazetteer list, not considered in the document when matching)
+            getterfunc: a callable which, given a token annotation, retrieves the string. If there is mapfunc, the
+                retrieved string is then still run through the mapfunc. The getterfunc must accept the token and
+                an optional document as parameters.
+            list_features: a dictionary of features common to the whole list loaded or None.
+                If what gets loaded specifies
+                its own list features, this is getting ignored.
+            list_type: the output annotation type to use for the list, ignored if the input format specifies this
+                on its own. If the input does not specify this on its own and this is not None, then it takes
+                precedence over outtype for the data loaded from source.
 
         """
         self.nodes = defaultdict(TokenGazetteerNode)
         self.mapfunc = mapfunc
         self.ignorefunc = ignorefunc
         self.feature = feature
-        self.annset = annset
-        self.tokentype = tokentype
-        self.septype = septype
-        self.splittype = splittype
-        self.withintype = withintype
-        self.outset = outset
-        self.outtype = outtype
-        self.all = all
-        self.skip = skip
-        self.tokenizer = tokenizer
+        self.annset = annset_name
+        self.tokentype = token_type
+        self.splittype = split_type
+        self.withintype = within_type
+        self.outset = outset_name
+        self.outtype = ann_type
+        self.longest_only = longest_only
+        self.skip = skip_longest
         if getterfunc:
             self.getterfunc = getterfunc
         else:
@@ -158,20 +165,21 @@ class TokenGazetteer(GazetteerAnnotator):
         self.listtypes = []
         self.logger = init_logger(__name__)
         # self.logger.setLevel(logging.DEBUG)
-        self.source_sep = source_sep
-        self.source_encoding = source_encoding
-        self.append(source, fmt=fmt, listfeatures=listfeatures, listtype=listtype, source_sep=source_sep,
-                    source_encoding=source_encoding
-                    )
+        if source is not None:
+            self.append(source, source_fmt=source_fmt,
+                        list_features=list_features, list_type=list_type, source_sep=source_sep,
+                        source_encoding=source_encoding, source_tokenizer=source_tokenizer
+                        )
 
     def append(
         self,
-        source,
-        fmt="gate-def",
-        source_sep="\t",
-        source_encoding="UTF-8",
-        listfeatures=None,
-        listtype=None,
+        source:  Union[None, str, List],
+        source_fmt: str = "gate-def",
+        source_sep: str = "\t",
+        source_encoding: str = "UTF-8",
+        source_tokenizer: Union[None, Annotator, Callable] = None,
+        list_features: Optional[Dict] = None,
+        list_type: Optional[str] = None,
     ):
         """
         This method appends more entries to gazetteer.
@@ -179,25 +187,30 @@ class TokenGazetteer(GazetteerAnnotator):
         Args:
             source: where to load the gazetteer from. What is actually expected here depends on the fmt
               parameter.
-            fmt: defines what is expected as the format and/or content of the source parameter. One of:
+            source_fmt: defines what is expected as the format and/or content of the source parameter. One of:
                *  "gate-def" (default): source must be a string, a pathlib Path or a parsed urllib url and
                   point to a GATE-style "def" file. See https://gate.ac.uk/userguide/chap:gazetteers
                * "gazlist": a list of tuples or lists where the first element of the tuple/list
                   is a list of strings, the second element is a dictionary containing the features to assign and
                   the third element, if it exists, is the index of an element in the listfeatures array.
-            listfeatures: a list of dictionaries containing the features to set for all matches witch have the
+            source_sep: the field separator to use for some source formats (default: tab character)
+            source_encoding: the encoding to use for some source formats (default: UTF-8)
+            source_tokenizer: if not None, an annotator, that creates annotations of type "Token" in the default
+                annotation set. If this is None, then when loading string gazetteer entries, they are tokenized by
+                splitting on whitespace (as defined by Python str.split())
+            list_features: a list of dictionaries containing the features to set for all matches witch have the
               list index set, this list gets appended to the existing listfeatures. If what gets appended specifies
               its own list features, this is ignored.
-            listtype: the output annotation type to use for the list that gets appended. If what gets appended
+            list_type: the output annotation type to use for the list that gets appended. If what gets appended
                specifies its own list type or list types, this is ignored.
         """
-        if fmt == "gazlist":
-            if listfeatures is not None:
-                self.listfeatures.append(listfeatures)
+        if source_fmt == "gazlist":
+            if list_features is not None:
+                self.listfeatures.append(list_features)
             else:
                 self.listfeatures.append({})
-            if listtype is not None:
-                self.listtypes.append(listtype)
+            if list_type is not None:
+                self.listtypes.append(list_type)
             else:
                 self.listtypes.append(self.outtype)
             listidx = len(self.listfeatures) - 1
@@ -205,11 +218,11 @@ class TokenGazetteer(GazetteerAnnotator):
                 entry = el[0]
                 data = el[1]
                 self.add(entry, data, listidx=listidx)
-        elif fmt == "gate-def":
-            if listfeatures is None:
-                listfeatures = {}
-            if listtype is None:
-                listtype = self.outtype
+        elif source_fmt == "gate-def":
+            if list_features is None:
+                list_features = {}
+            if list_type is None:
+                list_type = self.outtype
             with open(source, "rt", encoding=source_encoding) as infp:
                 for line in infp:
                     line = line.rstrip("\n\r")
@@ -220,8 +233,8 @@ class TokenGazetteer(GazetteerAnnotator):
                     minorType = fields[2]
                     languages = fields[3]
                     anntype = fields[4]
-                    this_listfeatures = listfeatures.copy()
-                    this_outtype = listtype
+                    this_listfeatures = list_features.copy()
+                    this_outtype = list_type
                     if majorType:
                         this_listfeatures["majorType"] = majorType
                     if minorType:
@@ -242,13 +255,9 @@ class TokenGazetteer(GazetteerAnnotator):
                             listline = listline.rstrip("\n\r")
                             fields = listline.split(source_sep)
                             entry = fields[0]
-                            if self.tokenizer:
+                            if source_tokenizer:
                                 tmpdoc = Document(entry)
-                                tmpdoc = self.tokenizer(tmpdoc)  # we MUST reassign here to allow return of a new doc!
-                                # TODO: include and handle SpaceToken if we use the speparator annoations!
-                                # TODO: maybe have a different way to retrieve the token annotations based
-                                # on the tokenizer????
-                                # TODO: try to figure out the outset and type used by the tokenizer, for now fixed! 
+                                tmpdoc = source_tokenizer(tmpdoc)  # we MUST reassign here to allow return of a new doc!
                                 tokenanns = list(tmpdoc.annset().with_type("Token"))
                                 if self.getterfunc:
                                     tokenstrings = [
@@ -284,7 +293,7 @@ class TokenGazetteer(GazetteerAnnotator):
                             listidx = len(self.listfeatures) - 1
                             self.add(tokenstrings, feats, listidx=listidx)
         else:
-            raise Exception(f"TokenGazetteer format {fmt} not known")
+            raise Exception(f"TokenGazetteer format {source_fmt} not known")
 
     def add(self, entry, data=None, listidx=None):
         """
@@ -300,9 +309,6 @@ class TokenGazetteer(GazetteerAnnotator):
                represents a token to be matched
             data: dictionary of features to add
             listidx: the index to list features and a list type to add
-
-        Returns:
-            The list of feature sets currently stored with the entry
         """
         if isinstance(entry, str):
             entry = [entry]
@@ -374,9 +380,8 @@ class TokenGazetteer(GazetteerAnnotator):
         #     else:
         #         node.listidx.append(None)
         #         node.listidx.append(listidx)
-        return node.data, node.listidx
 
-    def match(self, tokens, doc=None, all=None, idx=0, endidx=None, matchfunc=None):
+    def match(self, tokens, doc=None, longest_only=None, idx=0, endidx=None, matchfunc=None):
         """
         Try to match at index location idx of the tokens sequence. Returns a list which contains
         no elements if no match is found,  or
@@ -388,7 +393,7 @@ class TokenGazetteer(GazetteerAnnotator):
             tokens: a list of tokens (must allow to fetch the ith token as tokens[i])
             doc: the document to which the tokens belong. Necessary of the underlying text is used
                for the tokens.
-            all: whether to return all matches or just the longest ones. If None, overrides the setting
+            longest_only: whether to return all matches or just the longest ones. If None, overrides the setting
                from init.
             idx: the index in tokens where the match must start
             endidx: the index in tokens after which no match must end
@@ -404,8 +409,8 @@ class TokenGazetteer(GazetteerAnnotator):
         if endidx is None:
             endidx = len(tokens)
         assert idx < endidx
-        if all is None:
-            all = self.all
+        if longest_only is None:
+            longest_only = self.longest_only
         token = tokens[idx]
         if token.type == self.splittype:
             return [], 0
@@ -477,11 +482,11 @@ class TokenGazetteer(GazetteerAnnotator):
                                     node.data,
                                     node.listidx,
                                 )
-                            debugtxt = " ".join(
-                                [doc[tokens[i]] for i in range(match.start, match.end)]
-                            )
+                            # debugtxt = " ".join(
+                            #     [doc[tokens[i]] for i in range(match.start, match.end)]
+                            # )
                             # TODO: should LONGEST get calculated including ignored tokens or not?
-                            if all:
+                            if not longest_only:
                                 thismatches.append(match)
                                 if len(thistokens) > longest:
                                     longest = len(thistokens)
@@ -502,13 +507,13 @@ class TokenGazetteer(GazetteerAnnotator):
 
     def find(
         self,
-        tokens,
-        doc=None,
-        all=None,
-        fromidx=None,
-        toidx=None,
-        endidx=None,
-        matchfunc=None,
+        tokens: List[Annotation],
+        doc: Optional[Document] = None,
+        longest_only: Optional[bool] = None,
+        fromidx: Optional[int] = None,
+        toidx: Optional[int] = None,
+        endidx: Optional[int] = None,
+        matchfunc: Optional[Callable] = None,
     ):
         """
         Find the next match in the given index range and return a tuple with two elements: the first element
@@ -519,7 +524,7 @@ class TokenGazetteer(GazetteerAnnotator):
             tokens: list of tokens (must allow to fetch the ith token as tokens[i])
             doc: the document to which the tokens belong. Necessary of the underlying text is used
                for the tokens.
-            all: whether to return all matches or just the longest ones. If not none, overrides the
+            longest_only: whether to return all matches or just the longest ones. If not none, overrides the
                setting from init
             fromidx: first index where a match may start
             toidx: last index where a match may start
@@ -531,8 +536,8 @@ class TokenGazetteer(GazetteerAnnotator):
             as the second element and the index where the match occurs or None as the third element
 
         """
-        if all is None:
-            all = self.all
+        if longest_only is None:
+            longest_only = self.longest_only
         idx = fromidx
         if toidx is None:
             toidx = len(tokens) - 1
@@ -540,7 +545,7 @@ class TokenGazetteer(GazetteerAnnotator):
             endidx = len(tokens)
         while idx <= toidx:
             matches, long = self.match(
-                tokens, idx=idx, doc=doc, all=all, endidx=endidx, matchfunc=matchfunc
+                tokens, idx=idx, doc=doc, longest_only=longest_only, endidx=endidx, matchfunc=matchfunc
             )
             if long == 0:
                 idx += 1
@@ -550,14 +555,14 @@ class TokenGazetteer(GazetteerAnnotator):
 
     def find_all(
         self,
-        tokens,
-        doc=None,
-        all=None,
-        skip=None,
-        fromidx=None,
-        toidx=None,
-        endidx=None,
-        matchfunc=None,
+        tokens: List[Annotation],
+        doc: Optional[Document] = None,
+        longest_only: Optional[bool] = None,
+        skip_longest: Optional[bool] = None,
+        fromidx: Optional[int] = None,
+        toidx: Optional[int] = None,
+        endidx: Optional[int] = None,
+        matchfunc: Optional[Callable] = None,
         # reverse=True,
     ):
         """
@@ -570,9 +575,9 @@ class TokenGazetteer(GazetteerAnnotator):
                string.
             doc: the document this should run on. Only necessary if the text to match is not retrieved from
                the token annotation, but from the underlying document text.
-            all: return all matches, if False only return longest matches. If not None, overrides the init
+            longest_only: whether to return only the longest or all matches. If not None, overrides the init
                setting
-            skip: skip forward over longest match (do not return contained/overlapping matches). If not
+            skip_longest: skip forward over longest match (do not return contained/overlapping matches). If not
                None overrides the init setting.
             fromidx: index where to start finding in tokens
             toidx: index where to stop finding in tokens (this is the last index actually used)
@@ -584,10 +589,10 @@ class TokenGazetteer(GazetteerAnnotator):
             list of matches
         """
         logger = self.logger
-        if all is None:
-            all = self.all
-        if skip is None:
-            skip = self.skip
+        if longest_only is None:
+            longest_only = self.longest_only
+        if skip_longest is None:
+            skip_longest = self.skip
         matches = []
         lentok = len(tokens)
         if endidx is None:
@@ -609,7 +614,7 @@ class TokenGazetteer(GazetteerAnnotator):
             matches, maxlen, idx = self.find(
                 tokens,
                 doc=doc,
-                all=all,
+                longest_only=longest_only,
                 fromidx=idx,
                 endidx=endidx,
                 toidx=toidx,
@@ -618,70 +623,38 @@ class TokenGazetteer(GazetteerAnnotator):
             if idx is None:
                 return
             yield matches
-            if skip:
+            if skip_longest:
                 idx += maxlen
             else:
                 idx += 1
 
     def __call__(
         self,
-        doc,
-        annset=None,
-        tokentype=None,
-        septype=None,
-        splittype=None,
-        withintype=None,
-        all=None,
-        skip=None,
+        doc: Document,
     ):
         """
         Apply the gazetteer to the document and annotate all matches.
 
         Args:
             doc: the document to annotate with matches.
-            annset: if given, overrides the one specified for the gazetteer instance.
-            tokentype: if given, overrides the one specified for the gazetteer instance.
-            septype: if given, overrides the one specified for the gazetteer instance.
-            splittype: if given, overrides the one specified for the gazetteer instance.
-            withintype: if given, overrides the one specified for the gazetteer instance.
-            all: if not None, overrides the gazetteer setting
-            skip: if not None, overrides the gazetteer setting
-
 
         Returns:
             the annotated document
         """
-        if all is None:
-            all = self.all
-        if skip is None:
-            skip = self.skip
-        if annset is None:
-            annset = self.annset
-        if tokentype is None:
-            tokentype = self.tokentype
-        if septype is None:
-            septype = self.septype
-        if splittype is None:
-            splittype = self.splittype
-        if withintype is None:
-            withintype = self.withintype
         # create the token lists from the document: if withintype is None we only have one token list,
         # otherwise we have one list for each withingtype
         # We create a list of segments which are identified by start and end offsets
-        if withintype is None:
+        if self.withintype is None:
             segment_offs = [(0, len(doc.text))]
         else:
-            withinanns = doc.annset(withintype)
+            withinanns = doc.annset(self.withintype)
             segment_offs = []
             for wann in withinanns:
                 segment_offs.append((wann.start, wann.end))
-        anntypes = [tokentype]
-        # TODO: septype still not implemented, not added here!
-        # if septype is not None:
-        #     anntypes.append(septype)
-        if splittype is not None:
-            anntypes.append(splittype)
-        anns = doc.annset(annset).with_type(anntypes)
+        anntypes = [self.tokentype]
+        if self.splittype is not None:
+            anntypes.append(self.splittype)
+        anns = doc.annset(self.annset).with_type(anntypes)
         # now do the annotation process for each segment
         outset = doc.annset(self.outset)
         for segment_start, segment_end in segment_offs: